Combine all little LoadingPlaceholderV2 components into one file
